docs(ci): extract injectSection to utils to avoid ESM cycle; update generator
